+++ planning/autoware_trajectory_validator/include/autoware/trajectory_validator/detail/validator_context.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__TRAJECTORY_VALIDATOR__EVALUATION_CONTEXT_HPP_
-#define AUTOWARE__TRAJECTORY_VALIDATOR__EVALUATION_CONTEXT_HPP_
+#ifndef AUTOWARE__TRAJECTORY_VALIDATOR__VALIDATOR_CONTEXT_HPP_
+#define AUTOWARE__TRAJECTORY_VALIDATOR__VALIDATOR_CONTEXT_HPP_
 
 #include <autoware_perception_msgs/msg/predicted_objects.hpp>
 #include <autoware_perception_msgs/msg/traffic_light_group_array.hpp>
@@ -31,7 +31,7 @@ namespace autoware::trajectory_validator
  * @brief A read-only snapshot of the world state for a single validation stage execution.
  * Callees receive it by `const &` and must not mutate fields.
  */
-struct EvaluationContext
+struct ValidatorContext
 {
   nav_msgs::msg::Odometry::ConstSharedPtr odometry;
   geometry_msgs::msg::AccelWithCovarianceStamped::ConstSharedPtr acceleration;
@@ -40,9 +40,7 @@ struct EvaluationContext
   autoware_perception_msgs::msg::TrafficLightGroupArray::ConstSharedPtr traffic_light_signals;
 };
 
-// Alias to maintain backward compatibility with existing validator plugins
-using FilterContext = EvaluationContext;
-
+using FilterContext = ValidatorContext;
 }  // namespace autoware::trajectory_validator
 
-#endif  // AUTOWARE__TRAJECTORY_VALIDATOR__EVALUATION_CONTEXT_HPP_
+#endif  // AUTOWARE__TRAJECTORY_VALIDATOR__VALIDATOR_CONTEXT_HPP_
