+++ planning/autoware_trajectory_validator/include/autoware/trajectory_validator/trajectory_validator_interface.hpp
@@ -16,11 +16,10 @@
 #define AUTOWARE__TRAJECTORY_VALIDATOR__TRAJECTORY_VALIDATOR_INTERFACE_HPP_
 
 #include "autoware/trajectory_validator/detail/trajectory_validator.hpp"
-#include "autoware/trajectory_validator/evaluation_context.hpp"
-#include "autoware/trajectory_validator/trajectory_validator_report.hpp"
+#include "autoware/trajectory_validator/detail/trajectory_validator_report.hpp"
+#include "autoware/trajectory_validator/detail/validator_context.hpp"
 #include "autoware/trajectory_validator/validator_interface.hpp"
 
-#include <autoware/trajectory_validator/parameters.hpp>
 #include <autoware_trajectory_validator/msg/metric_report.hpp>
 #include <autoware_trajectory_validator/msg/validation_report.hpp>
 #include <autoware_trajectory_validator/msg/validation_report_array.hpp>
@@ -69,7 +68,7 @@ public:
     std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper);
 
   CandidateTrajectories validate_trajectories(
-    const CandidateTrajectories & input_trajectories, const EvaluationContext & context);
+    const CandidateTrajectories & input_trajectories, const ValidatorContext & context);
 
 private:
   void publishers();
