+++ planning/autoware_trajectory_validator/include/autoware/trajectory_validator/validator_interface.hpp
@@ -15,7 +15,7 @@
 #ifndef AUTOWARE__TRAJECTORY_VALIDATOR__VALIDATOR_INTERFACE_HPP_
 #define AUTOWARE__TRAJECTORY_VALIDATOR__VALIDATOR_INTERFACE_HPP_
 
-#include "autoware/trajectory_validator/filter_context.hpp"
+#include "autoware/trajectory_validator/detail/validator_context.hpp"
 
 #include <autoware_trajectory_validator/autoware_trajectory_validator_param.hpp>
 #include <autoware_trajectory_validator/msg/metric_report.hpp>
